refactor list components table columns into hooks
