+++ src/hooks/use-entitlement-table-columns.tsx
@@ -1,0 +1,37 @@
+import { useMemo } from "react"
+
+import { Entitlement } from "@/types/entitlements"
+import { createTableColumnHelper } from "@/lib/tables"
+import ClipboardButton from "@/components/clipboard-button"
+
+const column = createTableColumnHelper<Entitlement>()
+
+export function useEntitlementTableColumns() {
+  const columns = useMemo(
+    () => [
+      column.id({
+        header: "ID",
+        cell: (info) => <ClipboardButton value={info.getValue()} />,
+      }),
+      column.attr("name", {
+        header: "Name",
+      }),
+      column.attr("code", {
+        header: "Code",
+      }),
+      column.attr("created", {
+        sortingFn: "datetime",
+        header: "Created",
+        cell: (info) => new Date(info.getValue()).toLocaleDateString(),
+      }),
+      column.attr("updated", {
+        sortingFn: "datetime",
+        header: "Updated",
+        cell: (info) => new Date(info.getValue()).toLocaleDateString(),
+      }),
+    ],
+    [],
+  )
+
+  return columns
+}
