+++ src/pages/app/product/list.tsx
@@ -1,10 +1,11 @@
-import { useMemo, useState } from "react"
+import { useState } from "react"
 import { useNavigate } from "@tanstack/react-router"
 
 import { Button } from "@/components/ui/button"
 import { Dialog, DialogTrigger } from "@/components/ui/dialog"
 
-import { createTableColumnHelper } from "@/lib/tables"
+import { useProductTableColumns } from "@/hooks/use-product-table-columns"
+
 import { Product } from "@/types/products"
 
 import { useListProducts } from "@/queries/products"
@@ -14,7 +15,6 @@ import * as Products from "@/components/products"
 import DataTable from "@/components/data-table"
 import PageHeader from "@/components/page-header"
 import SkeletonTable from "@/components/skeleton-table"
-import ClipboardButton from "@/components/clipboard-button"
 
 export default function ProductsList() {
   const { data: products = [], isLoading } = useListProducts()
@@ -22,45 +22,7 @@ export default function ProductsList() {
 
   const [open, setOpen] = useState(false)
 
-  const column = createTableColumnHelper<Product>()
-  const columns = useMemo(
-    () => [
-      column.id({
-        header: "ID",
-        cell: (info) => <ClipboardButton value={info.getValue()} />,
-      }),
-      column.attr("name", {
-        header: "Name",
-      }),
-      column.attr("code", {
-        header: "Code",
-      }),
-      column.attr("url", {
-        header: "URL",
-        cell: (info) => (
-          <a
-            href={info.getValue()}
-            target="_blank"
-            className="text-primary hover:underline"
-            onClick={(e) => e.stopPropagation()}
-          >
-            {info.getValue()}
-          </a>
-        ),
-      }),
-      column.attr("created", {
-        header: "Created",
-        cell: (info) => new Date(info.getValue()).toLocaleDateString(),
-        sortingFn: "datetime",
-      }),
-      column.attr("updated", {
-        header: "Updated",
-        cell: (info) => new Date(info.getValue()).toLocaleDateString(),
-        sortingFn: "datetime",
-      }),
-    ],
-    [],
-  )
+  const columns = useProductTableColumns()
 
   const handleSelectProduct = (product: Product | null) => {
     if (!product) return
